hoptix-flask/scripts/test_timing_only.py
#!/usr/bin/env python3
"""
Test script to verify timestamp parsing and transaction timing logic without requiring video files.
"""

import os
import sys
from datetime import datetime, timedelta

# Add parent directory to path
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))

from worker.adapter import _parse_dt_file_timestamp, _iso_from_start

def test_timing_logic():
    """Test the timing parsing and calculation logic"""
    
    print("🧪 Testing Transaction Timing Logic")
    print("=" * 60)
    
    # Test 1: DT_File timestamp parsing
    print("\n📅 Test 1: DT_File Timestamp Parsing")
    test_cases = [
        "dev/sample/DT_File20250817210001000.avi",
        "dev/sample/DT_File20250824120530500.avi", 
        "dev/sample/DT_File20250301083045123.avi",
        "dev/sample/normal_file.avi"  # Should fallback
    ]
    
    for s3_key in test_cases:
        parsed = _parse_dt_file_timestamp(s3_key)
        print(f"   {s3_key}")
        print(f"   → {parsed}")
    
    # Test 2: Transaction timing calculation
    print("\n⏱️  Test 2: Transaction Timing Calculation")
    
    # Simulate a video starting at a specific time
    video_start = "2025-08-17T21:00:01Z"
    print(f"   Video starts at: {video_start}")
    
    # Simulate transaction segments within the video
    transactions = [
        {"start_seconds": 5.2, "end_seconds": 18.7, "description": "First order - Burger meal"},
        {"start_seconds": 25.0, "end_seconds": 42.3, "description": "Second order - Coffee and donuts"},
        {"start_seconds": 50.1, "end_seconds": 63.8, "description": "Third order - Chicken nuggets"}
    ]
    
    print("   Calculated transaction timestamps:")
    for i, tx in enumerate(transactions, 1):
        start_time = _iso_from_start(video_start, tx["start_seconds"])
        end_time = _iso_from_start(video_start, tx["end_seconds"])
        duration = tx["end_seconds"] - tx["start_seconds"]
        
        print(f"   Transaction {i}: {tx['description']}")
        print(f"     Video seconds: {tx['start_seconds']:.1f} - {tx['end_seconds']:.1f} ({duration:.1f}s)")
        print(f"     Real time: {start_time} to {end_time}")
        print()
    
    # Test 3: CSV structure simulation
    print("📊 Test 3: CSV Data Structure")
    
    mock_transaction = {
        "video_id": "test-video-123",
        "run_id": "test-run-456", 
        "started_at": "2025-08-17T21:00:06.2Z",
        "ended_at": "2025-08-17T21:00:19.7Z",
        "tx_range": '["2025-08-17T21:00:06.2Z","2025-08-17T21:00:19.7Z")',
        "kind": "order",
        "meta": {
            "text": "Welcome to Burger King, can I take your order?",
            "complete_order": 1,
            "mobile_order": 0,
            "coupon_used": 0,
            "asked_more_time": 0,
            "out_of_stock_items": "0",
            "video_start_seconds": 5.2,
            "video_end_seconds": 18.7,
            "s3_key": "dev/sample/DT_File20250817210001000.avi",
            "segment_index": 0,
            "total_segments_in_video": 3
        }
    }
    
    print("   Sample transaction data structure:")
    import json
    print(json.dumps(mock_transaction, indent=4))
    
    # Test 4: Verify timing accuracy  
    print("\n🎯 Test 4: Timing Accuracy Verification")
    
    # Parse the video filename
    s3_key = "dev/sample/DT_File20250817210001000.avi"
    filename_time = _parse_dt_file_timestamp(s3_key)
    
    # Add transaction offset
    transaction_start = _iso_from_start(filename_time, 5.2)
    transaction_end = _iso_from_start(filename_time, 18.7)
    
    print(f"   Video filename: {s3_key}")
    print(f"   Video start time: {filename_time}")
    print(f"   Transaction occurs at video seconds 5.2-18.7")
    print(f"   Transaction real time: {transaction_start} to {transaction_end}")
    
    # Calculate duration
    from dateutil import parser as dateparse
    start_dt = dateparse.isoparse(transaction_start)
    end_dt = dateparse.isoparse(transaction_end)
    duration = (end_dt - start_dt).total_seconds()
    
    print(f"   Duration: {duration} seconds ✅")
    
    print("\n" + "=" * 60)
    print("🎉 All timing logic tests completed successfully!")
    print("\n💡 Summary:")
    print("   ✅ DT_File timestamp parsing works")
    print("   ✅ Transaction timing calculation works") 
    print("   ✅ CSV data structure is complete")
    print("   ✅ Real-world timing accuracy verified")

if __name__ == "__main__":
    test_timing_logic()



